# docs/api-contract.yaml
openapi: 3.0.3
info:
  title: Hypatia Backend API
  description: >
    Contract between the Flutter frontend and the FastAPI backend for the
    Hypatia LLM wiki application. Covers health, Classes CRUD, file
    upload/serve/delete, wiki tree/page CRUD/search/export/lint/rebuild,
    chat WebSocket + history, task status, and class backup/import.
  version: "0.4.0"
servers:
  - url: http://127.0.0.1:8000
    description: >-
      Local dev backend. Port may vary 8000-8010 if 8000 is taken
      (see Task 1.9 auto-launch/port management).
paths:
...
  /api/classes/{class_id}/files:
    parameters:
      - $ref: "#/components/parameters/ClassId"
    get:
      summary: List all files in a Class
      operationId: listFiles
      responses:
        "200":
          description: Files in this Class
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileRead"
    post:
      summary: Upload one or more files to a Class
      operationId: uploadFiles
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
              required: [files]
      responses:
        "202":
          description: Files accepted; conversion/processing runs asynchronously
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileUploadResponse"
  /api/classes/{class_id}/files/{file_id}:
    parameters:
      - $ref: "#/components/parameters/ClassId"
      - $ref: "#/components/parameters/FileId"
    get:
      summary: Get a single file's status/details
      operationId: getFile
      responses:
        "200":
          description: File details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileRead"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete a file and trigger wiki cleanup
      operationId: deleteFile
      responses:
        "204":
          description: File deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /api/classes/{class_id}/files/{file_id}/raw:
    parameters:
      - $ref: "#/components/parameters/ClassId"
      - $ref: "#/components/parameters/FileId"
    get:
      summary: Serve the raw uploaded file
      operationId: getFileRaw
      responses:
        "200":
          description: Raw file content
          content:
            application/octet-stream: {}
        "404":
          $ref: "#/components/responses/NotFound"
  /api/classes/{class_id}/files/{file_id}/converted:
    parameters:
      - $ref: "#/components/parameters/ClassId"
      - $ref: "#/components/parameters/FileId"
    get:
      summary: Serve the converted markdown version
      operationId: getFileConverted
      responses:
        "200":
          description: Converted markdown content
          content:
            text/markdown: {}
        "404":
          $ref: "#/components/responses/NotFound"
  /api/classes/{class_id}/files/{file_id}/open:
    parameters:
      - $ref: "#/components/parameters/ClassId"
      - $ref: "#/components/parameters/FileId"
    get:
      summary: Deep-link to file with Location header for frontend navigation
      operationId: openFile
      responses:
        "200":
          description: File metadata with optional Location header
          headers:
            Location:
              schema:
                type: string
              description: URI for the file's converted content if available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileRead"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/classes/{class_id}/wiki/tree:
    parameters:
      - $ref: "#/components/parameters/ClassId"
    get:
      summary: Get the wiki page tree structure for the sidebar
      operationId: getWikiTree
      responses:
        "200":
          description: Flat list of wiki page summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WikiTreeNode"
  /api/classes/{class_id}/wiki/index:
    parameters:
      - $ref: "#/components/parameters/ClassId"
    get:
      summary: Get the wiki index page content
      operationId: getWikiIndex
      responses:
        "200":
          description: Index page content (empty string if no index exists)
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: string
                required: [content]
  /api/classes/{class_id}/wiki/pages/{page_path}:
    parameters:
      - $ref: "#/components/parameters/ClassId"
      - name: page_path
        in: path
        required: true
        schema:
          type: string
        description: Slash-separated relative path to the wiki page
    get:
      summary: Get a wiki page
      operationId: getWikiPage
      responses:
        "200":
          description: Wiki page content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WikiPageRead"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Update a wiki page (user edit)
      operationId: updateWikiPage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WikiPageUpdate"
      responses:
        "200":
          description: Updated wiki page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WikiPageRead"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/classes/{class_id}/wiki/search:
    parameters:
      - $ref: "#/components/parameters/ClassId"
    get:
      summary: Full-text search wiki pages
      operationId: searchWiki
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WikiSearchResult"
  /api/classes/{class_id}/wiki/export:
    parameters:
      - $ref: "#/components/parameters/ClassId"
    post:
      summary: Export all wiki pages as a ZIP
      operationId: exportWiki
      responses:
        "200":
          description: ZIP file containing wiki pages
          content:
            application/zip: {}
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/classes/{class_id}/wiki/lint:
    parameters:
      - $ref: "#/components/parameters/ClassId"
    post:
      summary: Run LLM-powered lint on wiki pages
      operationId: lintWiki
      responses:
        "200":
          description: Lint report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LintReport"
        "503":
          $ref: "#/components/responses/LLMUnavailable"
  /api/classes/{class_id}/wiki/rebuild:
    parameters:
      - $ref: "#/components/parameters/ClassId"
    post:
      summary: Rebuild the wiki (preview or confirm)
      operationId: rebuildWiki
      parameters:
        - name: confirm
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: >-
            Without confirm: rebuild preview. With confirm: task started.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/RebuildPreview"
                  - $ref: "#/components/schemas/TaskStarted"
        "503":
          $ref: "#/components/responses/LLMUnavailable"
...
  /api/tasks/{task_id}:
    parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get a task's status
      operationId: getTask
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStatus"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/tasks/{task_id}/cancel:
    parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
    post:
      summary: Cancel a running task
      operationId: cancelTask
      responses:
        "202":
          description: Cancellation requested
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    ClassId:
      name: class_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    FileId:
      name: file_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    LLMUnavailable:
      description: LLM provider is unavailable
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: LLM provider is unavailable
              code:
                type: string
                example: LLM_UNAVAILABLE
            required: [detail, code]
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
        code:
          type: string
      required: [detail, code]
...
    FileType:
      type: string
      enum: [pdf, docx, pptx, xlsx, image, video, audio, markdown, other]
    FileStatus:
      type: string
      enum: [pending, processing, ready, error]
    FileRead:
      type: object
      properties:
        id:
          type: string
          format: uuid
        class_id:
          type: string
          format: uuid
        original_filename:
          type: string
        file_type:
          $ref: "#/components/schemas/FileType"
        file_size_bytes:
          type: integer
        raw_path:
          type: string
        converted_path:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/FileStatus"
        error_message:
          type: string
          nullable: true
        metadata_json:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - class_id
        - original_filename
        - file_type
        - file_size_bytes
        - raw_path
        - status
        - created_at
        - updated_at
    FileUploadResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        class_id:
          type: string
          format: uuid
        original_filename:
          type: string
        status:
          $ref: "#/components/schemas/FileStatus"
      required: [id, class_id, original_filename, status]
    WikiCategory:
      type: string
      enum: [source-summary, concept, entity, synthesis, index, log]
    WikiPageRead:
      type: object
      properties:
        id:
          type: string
          format: uuid
        class_id:
          type: string
          format: uuid
        path:
          type: string
        title:
          type: string
        category:
          $ref: "#/components/schemas/WikiCategory"
        content:
          type: string
        source_file_ids:
          type: array
          items:
            type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - class_id
        - path
        - title
        - category
        - content
        - created_at
        - updated_at
    WikiPageUpdate:
      type: object
      properties:
        content:
          type: string
      required: [content]
    WikiTreeNode:
      type: object
      properties:
        id:
          type: string
          format: uuid
        path:
          type: string
        title:
          type: string
        category:
          $ref: "#/components/schemas/WikiCategory"
        updated_at:
          type: string
          format: date-time
      required: [id, path, title, category, updated_at]
    WikiSearchResult:
      type: object
      properties:
        page_id:
          type: string
          format: uuid
        path:
          type: string
        title:
          type: string
        snippet:
          type: string
        rank:
          type: number
      required: [page_id, path, title, snippet, rank]
    LintReport:
      type: object
      properties:
        success:
          type: boolean
        issues:
          type: array
          items:
            type: object
            properties:
              page_path:
                type: string
              severity:
                type: string
              message:
                type: string
            required: [page_path, severity, message]
        error:
          type: string
          nullable: true
      required: [success, issues]
    RebuildPreview:
      type: object
      properties:
        page_count:
          type: integer
        file_count:
          type: integer
        estimated_seconds:
          type: integer
      required: [page_count, file_count]
    TaskStarted:
      type: object
      properties:
        task_id:
          type: string
        status:
          type: string
          enum: [started]
      required: [task_id, status]
    TaskStatus:
      type: object
      properties:
        task_id:
          type: string
        operation:
          type: string
        class_id:
          type: string
        status:
          type: string
          enum: [running, completed, failed, cancelled]
        progress:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
        error:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      required: [task_id, operation, class_id, status, created_at]
...